Fix remote leader epoch initialization
ReplicaManager.applyLocalFollowersDelta skips calling partition.makeFollower for partitions
with remote leader because we want to make consumer-like fetches. The problem is that makeFollower
also sets the partition's leader epoch, so it remains unitialized at -1.

With negative epoch, UnifiedLog.analyzeAndValidateRecords skips the batches in a tigh loop
causing high CPU consumption and logs flooded with:

[2025-09-24 14:00:51,999] INFO [UnifiedLog partition=my-topic-0, dir=/home/fvaleri/Documents/kafka/build/test/server2/data] Skipping batch RecordBatch(magic=2, offsets=[5, 5], sequence=[1, 1], partitionLeaderEpoch=0, isTransactional=false, isControlBatch=false, compression=none, timestampType=CreateTime, crc=1079926212) from an origin of REPLICATION because its partition leader epoch 0 is higher than the replica's current leader epoch -1 (org.apache.kafka.storage.internals.log.UnifiedLog)

It can be reproduced having multiple brokers and triggering leader elections.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>